// src/types/entity-types.ts
import { EnabledDisabledState } from "./data-dictionary-types";

/**
 * 账号
 */
export interface AuthAccount {
  /**
   * id
   */
  id: number;
  /**
   * 用户
   */
  user?: Partial<OcUser>;
  /**
   * 提供者代号
   */
  providerCode: string;
  /**
   * 登录账号
   */
  login?: string;
  /**
   * 密码
   */
  password?: string;
  /**
   * Email
   */
  email?: string;
  /**
   * 手机号
   */
  mobile?: string;
  /**
   * 外部系统账号Id
   */
  externalAccountId?: string;
  /**
   * 详细信息
   */
  details?: Record<string, any>;
  /**
   * 创建时间
   */
  createdAt?: string;
  /**
   * 创建人
   */
  createdBy?: Partial<OcUser>;
  /**
   * 更新时间
   */
  updatedAt?: string;
  /**
   * 更新人
   */
  updatedBy?: Partial<OcUser>;
  /**
   * 删除时间
   */
  deletedAt?: string;
  /**
   * 删除人
   */
  deletedBy?: Partial<OcUser>;
}

/**
 * 用户
 */
export interface OcUser {
  /**
   * id
   */
  id: number;
  /**
   * 姓名
   */
  name: string;
  /**
   * 登录账号
   */
  login: string;
  /**
   * 密码
   */
  password?: string;
  /**
   * 是否隐藏
   */
  hidden: boolean;
  /**
   * 状态
   */
  state: EnabledDisabledState;
  /**
   * Email
   */
  email?: string;
  /**
   * 员工号
   */
  employeeCode?: string;
  /**
   * 手机号
   */
  mobile?: string;
  /**
   * 分机号
   */
  telephoneExt?: string;
  /**
   * 外部系统用户Id
   */
  externalId?: string;
  /**
   * 部门
   */
  department?: Partial<OcDepartment>;
  /**
   * 角色
   */
  roles?: any;
  /**
   * 账户
   */
  accounts?: any;
  /**
   * 创建时间
   */
  createdAt?: string;
  /**
   * 创建人
   */
  createdBy?: Partial<OcUser>;
  /**
   * 更新时间
   */
  updatedAt?: string;
  /**
   * 更新人
   */
  updatedBy?: Partial<OcUser>;
  /**
   * 删除时间
   */
  deletedAt?: string;
  /**
   * 删除人
   */
  deletedBy?: Partial<OcUser>;
}

/**
 * 部门
 */
export interface OcDepartment {
  /**
   * id
   */
  id: number;
  /**
   * 上级部门
   */
  parent?: Partial<OcDepartment>;
  /**
   * Code
   */
  code?: string;
  /**
   * 名称
   */
  name?: string;
  /**
   * 排序
   */
  orderNum: number;
  /**
   * 用户
   */
  users?: any;
  /**
   * 外部系统部门Id
   */
  externalId?: string;
  /**
   * 创建时间
   */
  createdAt?: string;
  /**
   * 创建人
   */
  createdBy?: Partial<OcUser>;
  /**
   * 更新时间
   */
  updatedAt?: string;
  /**
   * 更新人
   */
  updatedBy?: Partial<OcUser>;
  /**
   * 删除时间
   */
  deletedAt?: string;
  /**
   * 删除人
   */
  deletedBy?: Partial<OcUser>;
}